+++ ChangeLog.md
@@ -1,3 +1,11 @@
+* 112 [bug] godfryd
+
+    Fixed handling situation when IP address to Kea Control Agent has changed.
+    Till now Stork was not able to detect this and was still communicating to
+    the old address. Now it checks if address has changed and updates it
+    in the database.
+    (Gitlab #409)
+
 * 111 [bug] marcin
 
     Corrected presentation of the HA state in the dashboard and